Made all fields on book object nullable.
Added inputs to the interface function for library data.

Added fake response to library data implementation.

Added library data value to the starting screen state. It is nullable and starts as null by default.

If there is currently no library data then the UI should have an alternate look to display that information rather than the same screen with no content.

"Retrieving you library".
"Cannot get your library".
A timer that apologises if its taking a while - concurrent timer.
Maybe a loading animation of a book in the middle of the screen with it's pages being flipped.